requirements.txt
Flask
flask_socketio
pyserial
requests
